# static/api.yaml
openapi: 3.0.0
info:
  title: Book Management API
  description: An API for managing books, including adding, listing, searching, updating, and deleting books.
  version: 1.0.0

servers:
  - url: http://localhost:5000

paths:
  /books/add-book:
    post:
      summary: Add a new book
      description: Add a new book to the library with details such as Title, Author, Published Year, ISBN, and Genre.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - author
                - published_year
                - isbn
              properties:
                title:
                  type: string
                  description: The title of the book
                author:
                  type: string
                  description: The author of the book
                published_year:
                  type: integer
                  description: The year the book was published
                isbn:
                  type: integer
                  description: The ISBN of the book
                genre:
                  type: string
                  description: The genre of the book (optional)
      responses:
        '201':
          description: Book added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: The ID of the book
                  title:
                    type: string
                  author:
                    type: string
                  published_year:
                    type: integer
                  isbn:
                    type: integer
                  genre:
                    type: string
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /books/list-all-books:
    get:
      summary: List all books
      description: Retrieve a list of all books in the library.
      responses:
        '200':
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    author:
                      type: string
                    published_year:
                      type: integer
                    isbn:
                      type: integer
                    genre:
                      type: string
  /books/search-books:
    get:
      summary: Search for books
      description: Search for books by author, published year, or genre.
      parameters:
        - name: author
          in: query
          description: Filter books by author
          required: false
          schema:
            type: string
        - name: published_year
          in: query
          description: Filter books by published year
          required: false
          schema:
            type: integer
        - name: genre
          in: query
          description: Filter books by genre
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of filtered books
          content:
            application/json:
              schema:
                type: object
                properties:
                  Filtered books:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                        author:
                          type: string
                        published_year:
                          type: integer
                        isbn:
                          type: integer
                        genre:
                          type: string
  /books/{isbn}:
    put:
      summary: Update a book
      description: Update one or more details of a specific book using its ISBN.
      parameters:
        - name: isbn
          in: path
          required: true
          description: The ISBN of the book to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                author:
                  type: string
                published_year:
                  type: integer
                genre:
                  type: string
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  author:
                    type: string
                  published_year:
                    type: integer
                  isbn:
                    type: integer
                  genre:
                    type: string
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    delete:
      summary: Delete a book
      description: Remove a specific book from the library using its ISBN.
      parameters:
        - name: isbn
          in: path
          required: true
          description: The ISBN of the book to delete
          schema:
            type: integer
      responses:
        '204':
          description: Book deleted successfully
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string